fix: consolidate auth endpoints to comply with Vercel Hobby plan limit
