+++ vercel.json
@@ -1,5 +1,11 @@
 {
   "buildCommand": "pnpm build",
   "outputDirectory": "dist",
-  "framework": "vite"
+  "framework": "vite",
+  "rewrites": [
+    { "source": "/api/login", "destination": "/api/auth?action=login" },
+    { "source": "/api/signup", "destination": "/api/auth?action=signup" },
+    { "source": "/api/logout", "destination": "/api/auth?action=logout" },
+    { "source": "/api/me", "destination": "/api/auth?action=me" }
+  ]
 }
